Remove duplication between Rakefile and spec/support/connection
