+++ spec/support/connection.rb
@@ -1,14 +1,4 @@
 require "active_record"
-
-def db_config
-  {
-    adapter: "postgresql",
-    host:     ENV.fetch("PGHOST", "localhost"),
-    port:     ENV.fetch("PGPORT", "5432"),
-    database: ENV.fetch("TEST_DATABASE", "pg_enum_test"),
-    username: ENV.fetch("TEST_USER") { ENV.fetch("USER", "pg_enum") },
-    password: ENV["TEST_PASSWORD"]
-  }
-end
+require_relative "connection_config"
 
 ActiveRecord::Base.establish_connection(db_config)
